+++ app/src/main/java/dev/tp_94/mobileapp/self_made_cake/presentation/SelfMadeCake.kt
@@ -254,7 +254,7 @@ fun SelfMadeCakeStatelessScreen(
             DiameterSlider(
                 onChange = onDiameterChange,
                 diameter = state.cakeCustom.diameter,
-                valueRange = state.restrictions.minDiameter..state.restrictions.maxDiameter
+                valueRange = state.restrictions.minDiameter.toFloat()..state.restrictions.maxDiameter.toFloat()
             )
             Spacer(modifier = Modifier.height(8.dp))
             val expandedState = remember { mutableStateOf(false) }
